+++ blog.sh
@@ -8,6 +8,7 @@
 #   ./blog.sh unpublish [<slug>]
 #   ./blog.sh delete [<slug>]
 #   ./blog.sh restore [<slug>]
+#   ./blog.sh add --bundle --json  (base64 ZIP on stdin)
 #   ./blog.sh empty trash|versions
 #   ./blog.sh toot [<slug>]
 #   ./blog.sh rebuild [--full]